remove basic_def.h and change data type.

+++ modules/speech_enhance/include/SpeechEnhance.h
@@ -5,9 +5,9 @@
 extern "C" {
 #endif 
 
-#include "basic_def.h"
+#include <stdint.h>
 
-int32_t SpeechEnhance_Init(void** pHandle, uint16_t sample_rate, uint16_t nchannel, uint16_t fftlen, uint16_t nframe);
+int32_t SpeechEnhance_Init(void** pHandle, uint32_t sample_rate, uint32_t nchannel, uint32_t fftlen, uint32_t nframe);
 int32_t SpeechEnhance_Process(void* handle, int16_t* mic_inputs, int16_t* output);
 int32_t SpeechEnhance_Release(void* handle);
